MemoryPoolAllocator: fix reporting incorrect size.
During 'AddChunk', the amount allocated is the requested capacty plus the
size of a ChunkHeader. The additional size of the ChunkHeader is not taken
into account when the size is reported in the 'Size' method. If the
difference between the buffer size and the allocated size is less than the
size of a ChunkHeader, the user will allocate off the heap each time and
not know to increase the buffer's size.