CMake: use configure_package_config_file helper
The configure_package_config_file helper should be preferred when
creating the config file, to ensure the package is relocatable.

The current calculation of the relative path between CMAKE_INSTALL_DIR
and INCLUDE_INSTALL_DIR is not robust (they may be relative or absolute
path, and all cases should be considered).

The helper provides a macro that takes care of using the correct paths
relative to the install prefix of the package. The set_and_check macro
is also recommended to set variables inside the config file, to ensure
the paths exist.

When exporting the package to the CMake registry, the helper will just
use the full paths to the include directory in the sources.