Remove -march=native/-mcpu=native default compilation flags.
Indeed this is making the assumption that the build machine is also the
same (or in the case of -mcpu similar to) the CPU found on the run
machines. In the general cases of distributions where compiling happens
once and the file is binary is distributed on many other machines, this
is actually a wrong assumption.