Allow to parse only BigInt as String
Introduce kParseBigIntsAsStringsFlag

Some languages like Python or Ruby have arbitrary sized integers and
likely want to be able to parse them.

`kParseNumbersAsStringsFlag` allows this, but is much slower than
using the various `Int*` methods.

As such it is desirable to use `RawNumber` only when the number can't
otherwise be parsed.