Don't install file meant for use within local build
Remove the installation of RapidJSONConfig/ConfigVersion.cmake files
that were configured to the root of the build directory. According to
the comments, these are meant for somekind of usage within the local
build tree. The content of RapidJSONConfig.cmake refers to local
absolute directories, so installing it doesn't make sense.

Also, the previous attempts to install it resulted in a race with the
proper RapidJSONConfig.cmake file. This would be visible when running
the install target:

-- Installing: /.../install/lib/cmake/RapidJSON/RapidJSONConfig.cmake
-- Up-to-date: /.../install/lib/cmake/RapidJSON/RapidJSONConfig.cmake

In the worst case the file aimed for local environment would be
installed first resulting in a non-functioning install.

The "COMPONENT dev" is kept in the config file install for consistency
with other installation commands. As there is no EXCLUDE_FROM_ALL, the
"dev" component will be part of the default install target anyways.

Signed-off-by: Lasse Lopperi <lasse.lopperi@iki.fi>