Change an array to a vector
Member 'IterativeParsingReaderHandler::Logs' is not initialized in the
constructor and it's hard to initialize an array in the old version of
Vsiual Studio. Change it to a vector and remove the need for
LogCapacity and LogCount as separate members.

Move the const static variables out from the template struct
'IterativeParsingReaderHandler' to solve 'undefined reference for
LOG_*' link errors.